+++ sites/arolariu.ro/src/app/domains/invoices/view-invoice/[id]/island.tsx
@@ -71,7 +71,7 @@ export default function RenderViewInvoiceScreen({invoiceIdentifier}: Readonly<{i
             </span>
             <Link
               href={`/domains/invoices/edit-invoice/${id}`}
-              className='ml-auto flex rounded border-0 bg-indigo-500 px-6 py-2 text-white hover:bg-indigo-600 focus:outline-none'>
+              className='ml-auto flex rounded border-0 bg-indigo-600 px-6 py-2 text-white hover:bg-indigo-700 focus:outline-none'>
               Edit this invoice
             </Link>
             <button
