+++ sites/arolariu.ro/src/app/domains/page.tsx
@@ -18,7 +18,7 @@ export default function DomainsHomepage() {
     <main className='container mx-auto px-5 py-24'>
       <section className='flex flex-col'>
         <div className='h-1 overflow-hidden rounded bg-gray-200'>
-          <div className='h-full w-24 bg-indigo-500' />
+          <div className='h-full w-24 bg-indigo-700' />
         </div>
         <div className='mb-12 flex flex-col flex-wrap py-6 sm:flex-row'>
           <h1 className='align-center mb-2 justify-items-center bg-gradient-to-r from-pink-400 to-red-600 bg-clip-text text-center text-5xl font-bold text-transparent sm:mb-0 sm:w-2/5'>
